requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt
flask-wtf
flask-talisman
flask-limiter
python-dotenv
email_validator
pyotp
qrcode
bleach
